+++ test/threads_exec.jl
@@ -703,3 +703,86 @@ catch ex
     @test ex isa LoadError
     @test ex.error isa ArgumentError
 end
+
+@testset "@spawn interpolation" begin
+    # Issue #30896: evaluating argumentss immediately
+    begin
+        outs = zeros(5)
+        @sync begin
+            local i = 1
+            while i <= 5
+                Threads.@spawn setindex!(outs, $i, $i)
+                i += 1
+            end
+        end
+        @test outs == 1:5
+    end
+
+    # Args
+    @test fetch(Threads.@spawn 2+$2) == 4
+    @test fetch(Threads.@spawn Int($(2.0))) == 2
+    a = 2
+    @test fetch(Threads.@spawn *($a,$a)) == a^2
+    # kwargs
+    @test fetch(Threads.@spawn sort($([3 2; 1 0]), dims=2)) == [2 3; 0 1]
+    @test fetch(Threads.@spawn sort([3 $2; 1 $0]; dims=$2)) == [2 3; 0 1]
+
+    # Supports multiple levels of interpolation
+    @test fetch(Threads.@spawn "$($a)") == "$a"
+    let a = 1
+        # Interpolate the current value of `a` vs the value of `a` in the closure
+        t = Threads.@spawn :(+($$a, $a, a))
+        a = 2  # update `a` after spawning
+        @test fetch(t) == Expr(:call, :+, 1, 2, :a)
+    end
+
+    # Test the difference between different levels of interpolation
+    let
+        oneinterp  = Vector{Any}(undef, 5)
+        twointerps = Vector{Any}(undef, 5)
+        @sync begin
+            local i = 1
+            while i <= 5
+                Threads.@spawn setindex!(oneinterp, :($i), $i)
+                Threads.@spawn setindex!(twointerps, :($($i)), $i)
+                i += 1
+            end
+        end
+        # The first definition _didn't_ escape i
+        @test oneinterp == fill(6, 5)
+        # The second definition _did_ escape i
+        @test twointerps == 1:5
+    end
+end
+
+@testset "@async interpolation" begin
+    # Args
+    @test fetch(@async 2+$2) == 4
+    @test fetch(@async Int($(2.0))) == 2
+    a = 2
+    @test fetch(@async *($a,$a)) == a^2
+    # kwargs
+    @test fetch(@async sort($([3 2; 1 0]), dims=2)) == [2 3; 0 1]
+    @test fetch(@async sort([3 $2; 1 $0]; dims=$2)) == [2 3; 0 1]
+
+    # Supports multiple levels of interpolation
+    @test fetch(@async :($a)) == a
+    @test fetch(@async :($($a))) == a
+    @test fetch(@async "$($a)") == "$a"
+end
+
+# errors inside @threads
+function _atthreads_with_error(a, err)
+    Threads.@threads for i in eachindex(a)
+        if err
+            error("failed")
+        end
+        a[i] = Threads.threadid()
+    end
+    a
+end
+@test_throws TaskFailedException _atthreads_with_error(zeros(nthreads()), true)
+let a = zeros(nthreads())
+    _atthreads_with_error(a, false)
+    @test a == [1:nthreads();]
+end